Prevent log surging (#2)
This PR adds to `otel-api-common` a means for buffering logs, where logs
selected via a user-specified predicate will be buffered then later
flushed as aggregates, and all other logs will be logged immediately.

The motivation for these changes came from running
`opentelemetry-haskell` in production. When using the OTLP span exporter
and the observability backend (e.g. Honeycomb) is unavailable, there are
a few error states that can be encountered. Depending upon application
traffic, this can result in surges of error logging. Rapidly logging
every error, especially when the errors are often duplicates, is
inappropriate for most production applications. It's worth pointing out
that more surgical changes could've been made to `otel-sdk-trace`'s
`otlpSpanExporter`, but the approach taken in this PR offers some level
of future-proofing if users encounter other portions of
`opentelemetry-haskell` that could benefit from log buffering.

Using the `withBufferedLogger` function in this PR, users can granularly
control which of their loggers support buffering, and for those buffered
loggers, users can specify exactly which log messages should be
buffered.

An argument could be made that these changes should live somewhere else
(e.g. `monad-logger-aeson` or some other package), but it is convenient
to leverage `opentelemetry-haskell`'s `OnException` and `OnTimeout`
handlers in the buffered logger's config and I'd rather not move those
handlers out to some separate package at this time.

Note that it is not recommended to buffer all log messages. Buffering
error logs should be fine, but it is exceedingly unlikely that the guts
of the implementation here has appropriate throughput for buffering all
log messages under the sun.